requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
gitdb==4.0.12
gitpython==3.1.44
idna==3.10
jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
narwhals==1.46.0
numpy==1.26.4
packaging==24.2
pandas==2.3.1
pillow==10.4.0
protobuf==4.25.8
pyarrow==20.0.0
pydeck==0.9.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rich==13.9.4
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
streamlit==1.35.0
tenacity==8.5.0
toml==0.10.2
tornado==6.5.1
typing-extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
woocommerce==3.0.0
